Extract CLI serial reconnect runtime
